+++ between/src/main/java/com/wb/between/menu/repository/MenuRepository.java
@@ -9,7 +9,7 @@ import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface MenuRepository extends JpaRepository<Menu, BigInteger> {
+public interface MenuRepository extends JpaRepository<Menu, Long> {
 
     List<Menu> findByUseAt(String useAt, Sort sort);
 }
